requirements.txt
django
djangorestframework
# django-modeltranslation
ipdb
django-hvad